UBUNTU: (config) Change Xen paravirt drivers to be built-in
This changes CONFIG_XEN_PLATFORM_PCI, CONFIG_XEN_BLKDEV_FRONTEND and
CONFIG_XEN_NETDEV_FRONTEND to be built-in instead of modules.
The pci platform driver is required by the frontends and it was made
an unconditional built-in for 3.2 because of many mis-configurations.
The frontend drivers only take effect when a hypervisor (HVM or PVM)
is detected (and pull CONFIG_XENBUS_FRONTEND to be built-in). As the
PV drivers do not get autoloaded it is better to make them built-in.

BugLink: http://bugs.launchpad.net/bugs/886521

Signed-off-by: Alex Bligh <alex@alex.org.uk>
[v2 added modules.ignore to avoid FTBS]
Signed-off-by: Stefan Bader <smb@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Herton Krzesinski <herton.krzesinski@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>