Revert "block: export blk_{get,put}_queue()"
This reverts commit 9df626f8723c1edef5aeee6847d4bd7b52a7f4e0.

This is part of reverting the following block/scsi changes:
1. put stricter guards on queue dead checks
2. fix oops in scsi_run_queue()
3. block: add proper state guards to __elv_next_request
4. block: export blk_{get,put}_queue()
5. Fix oops caused by queue refcounting failure
6. fix crash in scsi_dispatch_cmd()

These patches were applied to Maverick 2.6.35.14 stable release. The
first patch is known to cause a regression, first reported with 2.6.38
kernel (Natty) at https://bugs.launchpad.net/bugs/793796, and later also
reported with 2.6.32 kernel when the same change went through an stable
update (https://bugs.launchpad.net/bugs/811745).

The next patches are fixes for other regressions introduced in the
first (with exception of patch 4, which is just a backport to allow
patch 5 to be used in older kernels). They are being reverted because
even with all these fixes, we still had reports of problems, as tested
by one reporter on bug 793796:
https://bugs.launchpad.net/ubuntu/+source/linux/+bug/793796/comments/49

In this comment, the kernel tested had all fixes above applied on a
2.6.38 release, and still failed. Due to this, and being a stable
release, they are being reverted for now (instead of pursuing or
applying more fixes on top), after the application of 2.6.35.14 stable
patches, to avoid same regressions also on Maverick.

This revert series is also being tracked on bug 802986.

BugLink: http://bugs.launchpad.net/bugs/898139
Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>